twitter/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-aiplatform
protobuf